Dynamic Sizing: The Prequel
dynamic sizing works! sort of. my thing wont build with your combo box
in there. if you want to test it just cut(or fix) the combo box xaml. It
is "functional" for now. Ill work more on making it "pretty" again soon.
also, i think we should ditch the progress bar and just have the slider,
but i left it in for now.